onepassword-provisioning/code/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
boto3==1.13.11
botocore==1.16.11
chardet==3.0.4
docutils==0.15.2
idna==2.9
itsdangerous==1.1.0
jmespath==0.10.0
multidict==4.7.6
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.14.0
slackclient==2.5.0
urllib3==1.25.9
yarl==1.4.2
